Move exchange calendars to their own directory
