+++ pandas_market_calendars/calendar_registry.py
@@ -1,32 +1,18 @@
 from .market_calendar import MarketCalendar
-from .exchange_calendar_asx import ASXExchangeCalendar
-from .exchange_calendar_bmf import BMFExchangeCalendar
-from .exchange_calendar_cboe import CFEExchangeCalendar
-from .exchange_calendar_cme import \
+from .exchange_calendars.asx import ASXExchangeCalendar
+from .exchange_calendars.bmf import BMFExchangeCalendar
+from .exchange_calendars.cme import \
     CMEEquityExchangeCalendar, \
     CMEBondExchangeCalendar
-from .exchange_calendar_cme_globex_base import CMEGlobexBaseExchangeCalendar
-from .exchange_calendar_cme_globex_agriculture import CMEGlobexAgricultureExchangeCalendar
-from .exchange_calendar_cme_globex_crypto import CMEGlobexCryptoExchangeCalendar
-from .exchange_calendar_cme_globex_fx import CMEGlobexFXExchangeCalendar
-from .exchange_calendar_cme_globex_energy_and_metals import CMEGlobexEnergyAndMetalsExchangeCalendar
-from .exchange_calendar_cme_globex_equities import CMEGlobexEquitiesExchangeCalendar
-from .exchange_calendar_cme_globex_fixed_income import CMEGlobexFixedIncomeCalendar
-from .exchange_calendar_eurex import EUREXExchangeCalendar
-from .exchange_calendar_hkex import HKEXExchangeCalendar
-from .exchange_calendar_ice import ICEExchangeCalendar
-from .exchange_calendar_iex import IEXExchangeCalendar
-from .exchange_calendar_jpx import JPXExchangeCalendar
-from .exchange_calendar_lse import LSEExchangeCalendar
-from .exchange_calendar_nyse import NYSEExchangeCalendar
-from .exchange_calendar_ose import OSEExchangeCalendar
-from .exchange_calendar_sifma import SIFMAUSExchangeCalendar, SIFMAUKExchangeCalendar, SIFMAJPExchangeCalendar
-from .exchange_calendar_six import SIXExchangeCalendar
-from .exchange_calendar_sse import SSEExchangeCalendar
-from .exchange_calendar_tsx import TSXExchangeCalendar
-from .exchange_calendar_bse import BSEExchangeCalendar
-from .exchange_calendar_tase import TASEExchangeCalendar
-from .exchange_calendars_mirror import *
+from .exchange_calendars.cme_globex_crypto import CMEGlobexCryptoExchangeCalendar
+from .exchange_calendars.cme_globex_fx import CMEGlobexFXExchangeCalendar
+from .exchange_calendars.cme_globex_fixed_income import CMEGlobexFixedIncomeCalendar
+from .exchange_calendars.eurex import EUREXExchangeCalendar
+from .exchange_calendars.hkex import HKEXExchangeCalendar
+from .exchange_calendars.jpx import JPXExchangeCalendar
+from .exchange_calendars.ose import OSEExchangeCalendar
+from .exchange_calendars.six import SIXExchangeCalendar
+from .exchange_calendars.mirror import *
 
 
 def get_calendar(name, open_time=None, close_time=None) -> MarketCalendar:
