+++ pandas_market_calendars/exchange_calendars/asx.py
@@ -3,8 +3,8 @@ from datetime import time
 from pandas.tseries.holiday import AbstractHolidayCalendar, GoodFriday, EasterMonday
 from pytz import timezone
 
-from .holidays_oz import *
-from .market_calendar import MarketCalendar
+from pandas_market_calendars.holidays_oz import *
+from pandas_market_calendars.market_calendar import MarketCalendar
 
 AbstractHolidayCalendar.start_date = '2011-01-01'
