+++ pandas_market_calendars/exchange_calendars/cboe.py
@@ -6,10 +6,10 @@ from pytz import timezone
 from itertools import chain
 import pandas as pd
 
-from .holidays_us import (Christmas, USBlackFridayInOrAfter1993, USIndependenceDay, USMartinLutherKingJrAfter1998,
+from pandas_market_calendars.holidays_us import (Christmas, USBlackFridayInOrAfter1993, USIndependenceDay, USMartinLutherKingJrAfter1998,
                           USMemorialDay, USNewYearsDay, HurricaneSandyClosings, USNationalDaysofMourning,
                           USJuneteenthAfter2022)
-from .market_calendar import MarketCalendar
+from pandas_market_calendars.market_calendar import MarketCalendar
 
 
 # TODO: In pandas 2.0.3 this no longer works as the dt passed in is the entire matrix and not a single date
@@ -114,5 +114,3 @@ class CBOEIndexOptionsExchangeCalendar(CFEExchangeCalendar):
         "market_open": ((None, time(8, 30)),),
         "market_close": ((None, time(15, 15)),)
     }
-
-
