+++ pandas_market_calendars/exchange_calendars/cme_globex_agriculture.py
@@ -14,7 +14,7 @@
 # limitations under the License.
 
 from abc import ABC, abstractmethod
-from .exchange_calendar_cme_globex_base import CMEGlobexBaseExchangeCalendar
+from .cme_globex_base import CMEGlobexBaseExchangeCalendar
 
 from datetime import time
 from itertools import chain
@@ -23,10 +23,10 @@ from pandas import Timestamp
 from pandas.tseries.holiday import AbstractHolidayCalendar, GoodFriday, USLaborDay, USPresidentsDay, USThanksgivingDay
 from pytz import timezone
 
-from .holidays_us import (Christmas, ChristmasEveBefore1993, ChristmasEveInOrAfter1993, USBlackFridayInOrAfter1993,
-                          USIndependenceDay, USMartinLutherKingJrAfter1998, USMemorialDay, USJuneteenthAfter2022,
-                          USNationalDaysofMourning, USNewYearsDay)
-from .market_calendar import MarketCalendar
+from pandas_market_calendars.holidays_us import (Christmas, ChristmasEveBefore1993, ChristmasEveInOrAfter1993, USBlackFridayInOrAfter1993,
+                                                 USIndependenceDay, USMartinLutherKingJrAfter1998, USMemorialDay, USJuneteenthAfter2022,
+                                                 USNationalDaysofMourning, USNewYearsDay)
+from pandas_market_calendars.market_calendar import MarketCalendar
