+++ pandas_market_calendars/exchange_calendars/iex.py
@@ -1,6 +1,6 @@
 from datetime import time
 from itertools import chain
-from .exchange_calendar_nyse import NYSEExchangeCalendar 
+from .nyse import NYSEExchangeCalendar
 from pandas.tseries.holiday import AbstractHolidayCalendar
 from pytz import timezone
