+++ pandas_market_calendars/exchange_calendars/nyse.py
@@ -244,7 +244,7 @@ from pandas_market_calendars.holidays_nyse import (
     # 2018
     GeorgeHWBushDeath2018
 )
-from .market_calendar import MarketCalendar
+from pandas_market_calendars.market_calendar import MarketCalendar
 
 # Useful resources for making changes to this file:
 # http://www.nyse.com/pdfs/closings.pdf
