+++ pandas_market_calendars/exchange_calendars/six.py
@@ -4,7 +4,7 @@ from pandas.tseries.holiday import (AbstractHolidayCalendar, Day, Easter, Easter
                                     previous_friday)
 from pytz import timezone
 
-from .market_calendar import (FRIDAY, MONDAY, MarketCalendar, THURSDAY, TUESDAY, WEDNESDAY)
+from pandas_market_calendars.market_calendar import (FRIDAY, MONDAY, MarketCalendar, THURSDAY, TUESDAY, WEDNESDAY)
 
 # New Year's Eve
 NewYearsEve = Holiday(
