+++ pandas_market_calendars/exchange_calendars/tase.py
@@ -1,7 +1,7 @@
 from datetime import time
 from pandas import Timestamp
 from pytz import timezone
-from .market_calendar import MarketCalendar
+from pandas_market_calendars.market_calendar import MarketCalendar
 
 TASEClosedDay = [
     # 2019
