+++ tests/test_bse_calendar.py
@@ -3,7 +3,7 @@ import pytest
 import pandas as pd
 import pytz
 
-from pandas_market_calendars.exchange_calendar_bse import BSEExchangeCalendar, BSEClosedDay
+from pandas_market_calendars.exchange_calendars.bse import BSEExchangeCalendar, BSEClosedDay
 
 
 def test_time_zone():
