+++ tests/test_cme_agriculture_calendar.py
@@ -1,7 +1,7 @@
 import pandas as pd
 import pytz
 
-from pandas_market_calendars.exchange_calendar_cme import CMEAgricultureExchangeCalendar
+from pandas_market_calendars.exchange_calendars.cme import CMEAgricultureExchangeCalendar
 
 
 def test_time_zone():
