+++ tests/test_cme_bond_calendar.py
@@ -1,7 +1,7 @@
 import pandas as pd
 import pytz
 
-from pandas_market_calendars.exchange_calendar_cme import CMEBondExchangeCalendar
+from pandas_market_calendars.exchange_calendars.cme import CMEBondExchangeCalendar
 
 
 def test_time_zone():
