+++ tests/test_exchange_calendar_cme_globex_energy_and_metals.py
@@ -4,7 +4,7 @@ import pandas as pd
 from pandas.testing import assert_index_equal
 from pandas.tseries.offsets import CustomBusinessDay
 
-from pandas_market_calendars.exchange_calendar_cme_globex_energy_and_metals import CMEGlobexEnergyAndMetalsExchangeCalendar
+from pandas_market_calendars.exchange_calendars.cme_globex_energy_and_metals import CMEGlobexEnergyAndMetalsExchangeCalendar
