+++ tests/test_exchange_calendar_cme_globex_fx.py
@@ -3,7 +3,7 @@ import pytest
 import pytz
 from pandas.tseries.offsets import Day, Hour, Minute
 
-from pandas_market_calendars.exchange_calendar_cme_globex_fx import CMEGlobexFXExchangeCalendar
+from pandas_market_calendars.exchange_calendars.cme_globex_fx import CMEGlobexFXExchangeCalendar
 
 TZ = 'America/Chicago'
