+++ tests/test_ice_calendar.py
@@ -1,6 +1,6 @@
 import pandas as pd
 
-from pandas_market_calendars.exchange_calendar_ice import ICEExchangeCalendar
+from pandas_market_calendars.exchange_calendars.ice import ICEExchangeCalendar
 
 
 def test_test_name():
