+++ tests/test_sse_calendar.py
@@ -3,7 +3,7 @@ import datetime
 import pandas as pd
 import pytz
 
-from pandas_market_calendars.exchange_calendar_sse import SSEExchangeCalendar
+from pandas_market_calendars.exchange_calendars.sse import SSEExchangeCalendar
 from pandas_market_calendars.holidays_cn import all_holidays
 
 all_holidays = pd.DatetimeIndex(all_holidays)
