+++ tests/test_tsx_calendar.py
@@ -2,7 +2,7 @@ import pandas as pd
 import pytz
 from pandas.testing import assert_index_equal
 
-from pandas_market_calendars.exchange_calendar_tsx import TSXExchangeCalendar, VictoriaDay
+from pandas_market_calendars.exchange_calendars.tsx import TSXExchangeCalendar, VictoriaDay
 
 
 def test_time_zone():
